Implementación formulario

+++ ClassEntities/Graficador.cs
@@ -9,9 +9,9 @@ namespace ClassEntities
     public class Graficador
     {
         public Componente componente { get; set; }
-        public int x1 { get; set; }
-        public int y1 { get; set; }
-        public int x2 { get; set; }
-        public int y2 { get; set; }
+        public float x1 { get; set; }
+        public float y1 { get; set; }
+        public float x2 { get; set; }
+        public float y2 { get; set; }
     }
 }
